feat: bucketsort start + dOmega cleanup

+++ clique-solver/solve/solve_via_vc.cpp
@@ -11,15 +11,15 @@
 
 using namespace std;
 
-//TODO: store the MC solution for output
-//TODO: store complement graphs for each right neighbourhood
+
 //TODO: sort first n-d vertices in ordering by right degree
+//TODO: binary search for p
 //TODO: maybe optimise subgraph/complement graph construction
 //TODO: use better data reductions
 //TODO: make better use of lower and upper Bounds, eg. heuristical lower bound
+//TODO: store complement graphs for each right neighbourhood
 int SolverViaVC::solve_via_vc(Graph& G)
 {
-    //solution_complement_graph = Graph();
     #if DEBUG
         print_success("Starting MC solver using VC solver");
         size_t initial_N = G.N;
@@ -28,22 +28,20 @@ int SolverViaVC::solve_via_vc(Graph& G)
 
     if(G.N == 0){ return 0; }
 
-    //compute degeneracy ordering and degeneracy d of G
+    // compute degeneracy ordering and degeneracy d of G
     auto result = degeneracy_ordering_rN(G);
     degeneracy_ordering = move(result.first);
     std::vector<std::vector<Vertex*>> right_neighbourhoodsT = move(result.second);
     d = degeneracy(degeneracy_ordering, right_neighbourhoodsT);
-    //print_degeracy_ordering_and_rneighbourhoods(degeneracy_ordering, right_neighbourhoods, G);
 
     #if DEBUG
         cout << "Graph with N=" << initial_N << " and M=" << initial_M << " has degeneracy d=" << d << std::endl;
     #endif
-    //int i = 0;
-    //right_neighbourhoods.reserve(right_neighbourhoodsT.size());
+
+    // copy right neighbourhoods to rn struct //TODO: unnecessary copy
     for(auto r : right_neighbourhoodsT){
         vector<Vertex*> x = r;
         right_neighbourhoods.push_back({move(x), -1, {}});
-        //i++;
     }
     //check lower and upper bounds
     clique_UB = degeneracy_UB(d);
@@ -54,37 +52,14 @@ int SolverViaVC::solve_via_vc(Graph& G)
         cout << "Computed LB=" << clique_LB << ", UB=" << clique_UB << std::endl;
     #endif
 
-    //data reduction
-    //apply_k_core(G, clique_LB);
-
-    //test_graph_consistency(G); //TODO: remove debug
-
-    //TODO: Possibly need to exclude vertices from ordering which were excluded by k-core or compute max k-core before ordering
-
-    #if DEBUG
-        cout << "Applied Data Reductions. new_N=" << G.N << ", new_M=" << G.M << std::endl;
-    #endif
-
-    // for(int p = clique_UB; p >= clique_LB; p--)
-    // {
-    //     if(solve_via_vc_for_p(G, p))
-    //     {
-    //         return clique_UB - p;
-    //     }
-    // }
+    // terminate early
     if(clique_LB == clique_UB)
     {
-        //copy Vertex* vector to solution string vector
-        maximum_clique = std::vector<std::string>();
-        for(size_t i = 0; i < LB_maximum_clique.size(); ++i)
-        {
-            Vertex* v = LB_maximum_clique[i];
-            size_t v_id = LB_maximum_clique[i]->id;
-            maximum_clique.push_back(G.name_table[v_id]);
-        }
+        set_maximum_clique(LB_maximum_clique, G);
         return clique_LB;
     }
 
+    // solve for increasing values of p (clique core gap)
     for(int p = 0; p < clique_UB; p++) //TODO: check if this can be reduced
     {
         if(solve_via_vc_for_p(G, p, clique_LB))
@@ -95,45 +70,26 @@ int SolverViaVC::solve_via_vc(Graph& G)
     return clique_LB;
 }
 
-Graph get_complement_subgraph(Graph & G, std::vector<Vertex*>& induced_set, int LB)
-{
-    G.set_restore();
-    G.MM_induced_subgraph(induced_set);
-    //apply_k_core(G, LB);
-    Graph complement = G.complementary_graph(G);
-    //test_graph_consistency(complement); //TODO: remove debug
-    G.restore();
-    return complement;
-}
-
 //MARK: SOLVER FOR P
-bool SolverViaVC::solve_via_vc_for_p(Graph &G, size_t p, int LB)
+bool SolverViaVC::solve_via_vc_for_p_with_sorting(Graph &G, size_t p, int LB)
 {
     //get candidate set D = {vi ∈ V | i ≤ n − d, rdeg(vi) ≥ d − p}
     vector<Vertex*> D = get_candidate_set(p);
-    // std::cout << "N= " << G.N << ", d=" <<  d << ", p=" << p << ", D.size()=" << D.size() << std::endl;
-    // for(Vertex* v : D)
-    // {
-    //     std::cout << G.name_table[v->id] << ", #rN(v)=" << right_neighbourhoods[v->id].size() << std::endl;
-    // }
 
     for(size_t i = 0; i < D.size(); ++i)
     {
-        int vc_size = 0;
-        bool found = false;
-        Graph complement;
         // a) construct ¬G[Vi], where Vi is the right-neighborhood of vi
         Vertex* vi = D[i];
         size_t r_neighbourhood_size = right_neighbourhoods[vi->id].neigh.size();
         size_t vc_UB = r_neighbourhood_size + p - d + 1;
         if(r_neighbourhood_size > 0)
         {
-             if(right_neighbourhoods[vi->id].vc_size == -1){
-                complement = get_complement_subgraph(G, right_neighbourhoods[vi->id].neigh, LB);
-                //apply_k_core(complement, LB);
+            //if not already solved
+             if(right_neighbourhoods[vi->id].vc_size == -1)
+             {
+                Graph complement = get_complement_subgraph(G, right_neighbourhoods[vi->id].neigh, LB);
                 //complement.UB = vc_UB; //bound search tree
-                //vc_size = solve(complement);
-                vc_size = solve(complement);
+                int vc_size = solve(complement);
                 right_neighbourhoods[vi->id].vc_size = vc_size;
                 right_neighbourhoods[vi->id].sol = extract_maximum_clique_solution_from_rn(right_neighbourhoods[vi->id], complement);
                 right_neighbourhoods[vi->id].sol.push_back(complement.name_table[vi->id]);
@@ -142,37 +98,20 @@ bool SolverViaVC::solve_via_vc_for_p(Graph &G, size_t p, int LB)
         }
 
         //  b) if ¬G[Vi] has a vertex cover of size qi := |Vi| + p − d, return true
-        /*if( found && vc_size == (int) (r_neighbourhood_size + p - d))
-        {
-            if(right_neighbourhoods[vi->id].empty())
-            {
-                maximum_clique.push_back(G.name_table[vi->id]);
-            }
-            else
-            {
-                extract_maximum_clique_solution(complement, vi);
-            }
-            complement.delete_all();
-            return true;
-        }
-        complement.delete_all();
-        */
-       int target = r_neighbourhood_size + p - d;
-       int got = right_neighbourhoods[vi->id].vc_size;
-       if(right_neighbourhoods[vi->id].vc_size == (int) (r_neighbourhood_size + p - d)){
+        int target = r_neighbourhood_size + p - d;
+        int current_vc_size = right_neighbourhoods[vi->id].vc_size;
+        if(current_vc_size == target){
             if(right_neighbourhoods[vi->id].neigh.empty())
             {
                 maximum_clique.push_back(G.name_table[vi->id]);
             }
             else
             {
-                //extract_maximum_clique_solution(complement, vi);
                 maximum_clique = right_neighbourhoods[vi->id].sol;
             }
             //complement.delete_all();
             return true;
-       }
-       //complement.delete_all();
+        }
     }
 
     // if we can't find mc from right neighbourhoods, take remaining vertices in ordering
@@ -184,8 +123,7 @@ bool SolverViaVC::solve_via_vc_for_p(Graph &G, size_t p, int LB)
     if(Vf.size() > 0)
     {
         complement = get_complement_subgraph(G, Vf, LB);
-        //apply_k_core(complement, LB);
-        complement.UB = vc_UB; //bound vc search tree
+        //complement.UB = vc_UB; //bound vc search tree
         vc_size_r = solve(complement);
     }
 
@@ -196,7 +134,77 @@ bool SolverViaVC::solve_via_vc_for_p(Graph &G, size_t p, int LB)
         {
             extract_maximum_clique_solution(complement);
         }
-        complement.delete_all();
+        //complement.delete_all();
+        return true;
+    }
+    complement.delete_all();
+    return false;
+}
+
+bool SolverViaVC::solve_via_vc_for_p(Graph &G, size_t p, int LB)
+{
+    //get candidate set D = {vi ∈ V | i ≤ n − d, rdeg(vi) ≥ d − p}
+    vector<Vertex*> D = get_candidate_set(p);
+
+    for(size_t i = 0; i < D.size(); ++i)
+    {
+        // a) construct ¬G[Vi], where Vi is the right-neighborhood of vi
+        Vertex* vi = D[i];
+        size_t r_neighbourhood_size = right_neighbourhoods[vi->id].neigh.size();
+        size_t vc_UB = r_neighbourhood_size + p - d + 1;
+        if(r_neighbourhood_size > 0)
+        {
+            //if not already solved
+             if(right_neighbourhoods[vi->id].vc_size == -1)
+             {
+                Graph complement = get_complement_subgraph(G, right_neighbourhoods[vi->id].neigh, LB);
+                //complement.UB = vc_UB; //bound search tree
+                int vc_size = solve(complement);
+                right_neighbourhoods[vi->id].vc_size = vc_size;
+                right_neighbourhoods[vi->id].sol = extract_maximum_clique_solution_from_rn(right_neighbourhoods[vi->id], complement);
+                right_neighbourhoods[vi->id].sol.push_back(complement.name_table[vi->id]);
+                complement.delete_all();
+            }
+        }
+
+        //  b) if ¬G[Vi] has a vertex cover of size qi := |Vi| + p − d, return true
+        int target = r_neighbourhood_size + p - d;
+        int current_vc_size = right_neighbourhoods[vi->id].vc_size;
+        if(current_vc_size == target){
+            if(right_neighbourhoods[vi->id].neigh.empty())
+            {
+                maximum_clique.push_back(G.name_table[vi->id]);
+            }
+            else
+            {
+                maximum_clique = right_neighbourhoods[vi->id].sol;
+            }
+            //complement.delete_all();
+            return true;
+        }
+    }
+
+    // if we can't find mc from right neighbourhoods, take remaining vertices in ordering
+    // construct ¬G[Vf], where Vf = {vf , . . . , vn} and f := n − d + 1
+    auto Vf = get_remaining_set();
+    int vc_size_r = 0;
+    Graph complement;
+    size_t vc_UB = p;
+    if(Vf.size() > 0)
+    {
+        complement = get_complement_subgraph(G, Vf, LB);
+        //complement.UB = vc_UB; //bound vc search tree
+        vc_size_r = solve(complement);
+    }
+
+    //if G[Vf] has a vertex cover of size qf := p − 1, return true
+    if(vc_size_r == (int) p - 1)
+    {
+        if(Vf.size() > 0)
+        {
+            extract_maximum_clique_solution(complement);
+        }
+        //complement.delete_all();
         return true;
     }
     complement.delete_all();
@@ -225,6 +233,15 @@ vector<Vertex*> SolverViaVC::get_remaining_set()
         R.push_back(degeneracy_ordering[i]);
     }
     return R;
+}
+
+Graph SolverViaVC::get_complement_subgraph(Graph & G, std::vector<Vertex*>& induced_set, int LB)
+{
+    G.set_restore();
+    G.MM_induced_subgraph(induced_set);
+    Graph complement = G.complementary_graph(G);
+    G.restore();
+    return complement;
 }
 
 std::vector<std::string> SolverViaVC::extract_maximum_clique_solution_from_rn(rn& right, Graph& complementGraph){
@@ -295,6 +312,17 @@ void SolverViaVC::extract_maximum_clique_solution(Graph& complementGraph, Vertex
 
 }
 
+void SolverViaVC::set_maximum_clique(std::vector<Vertex*>& vertices, Graph& G)
+{
+    maximum_clique = std::vector<std::string>();
+    for(size_t i = 0; i < vertices.size(); ++i)
+    {
+        size_t v_id = vertices[i]->id;
+        maximum_clique.push_back(G.name_table[v_id]);
+    }
+    return;
+}
+
 
 //----------------------- TESTING FUNCTIONS -----------------------
 
@@ -333,4 +361,13 @@ bool SolverViaVC::is_clique(std::vector<Vertex*> vertices)
         }
     }
     return true;
+}
+
+int Buckets::get_next()
+{
+    if(current_bucket == -1 && current_index == -1)
+    {
+        return -1;
+    }
+
 }
